Enable switch fps position on mobile device

+++ core/3d/CCBundle3D.cpp
@@ -298,7 +298,7 @@ bool Bundle3D::loadObj(MeshDatas& meshdatas,
             }
 
             // split into submesh according to material
-            std::map<int, std::vector<unsigned int>> subMeshMap;
+            std::map<int, std::vector<unsigned short>> subMeshMap;
             for (size_t k = 0, size = mesh.material_ids.size(); k < size; ++k)
             {
                 int id     = mesh.material_ids[k];
@@ -447,7 +447,7 @@ bool Bundle3D::loadMeshDatasBinary(MeshDatas& meshdatas)
 
         for (unsigned int k = 0; k < meshPartCount; ++k)
         {
-            std::vector<unsigned int> indexArray;
+            std::vector<unsigned short> indexArray;
             std::string meshPartid = _binaryReader.readString();
             meshData->subMeshIds.push_back(meshPartid);
             unsigned int nIndexCount;
@@ -599,7 +599,7 @@ bool Bundle3D::loadMeshDatasBinary_0_1(MeshDatas& meshdatas)
             return false;
         }
 
-        std::vector<unsigned int> indices;
+        std::vector<unsigned short> indices;
         indices.resize(nIndexCount);
         if (_binaryReader.read(&indices[0], 2, nIndexCount) != nIndexCount)
         {
@@ -720,7 +720,7 @@ bool Bundle3D::loadMeshDatasBinary_0_2(MeshDatas& meshdatas)
             return false;
         }
 
-        std::vector<unsigned int> indices;
+        std::vector<unsigned short> indices;
         indices.resize(nIndexCount);
         if (_binaryReader.read(&indices[0], 2, nIndexCount) != nIndexCount)
         {
@@ -777,14 +777,14 @@ bool Bundle3D::loadMeshDatasJson(MeshDatas& meshdatas)
         const rapidjson::Value& mesh_part_array = mesh_data[PARTS];
         for (rapidjson::SizeType i = 0, mesh_part_array_size = mesh_part_array.Size(); i < mesh_part_array_size; ++i)
         {
-            std::vector<unsigned int> indexArray;
+            std::vector<unsigned short> indexArray;
             const rapidjson::Value& mesh_part = mesh_part_array[i];
             meshData->subMeshIds.push_back(mesh_part[ID].GetString());
             // index_number
             const rapidjson::Value& indices_val_array = mesh_part[INDICES];
             for (rapidjson::SizeType j = 0, indices_val_array_size = indices_val_array.Size();
                  j < indices_val_array_size; ++j)
-                indexArray.push_back((unsigned int)indices_val_array[j].GetUint());
+                indexArray.push_back((unsigned short)indices_val_array[j].GetUint());
 
             meshData->subMeshIndices.push_back(indexArray);
             meshData->numIndex = (int)meshData->subMeshIndices.size();
@@ -1185,12 +1185,12 @@ bool Bundle3D::loadMeshDataJson_0_1(MeshDatas& meshdatas)
     unsigned int indexnum = mesh_data_body_array_0[INDEXNUM].GetUint();
 
     // indices
-    std::vector<unsigned int> indices;
+    std::vector<unsigned short> indices;
     indices.resize(indexnum);
 
     const rapidjson::Value& indices_val_array = mesh_data_body_array_0[INDICES];
     for (rapidjson::SizeType i = 0; i < indices_val_array.Size(); ++i)
-        indices[i] = (unsigned int)indices_val_array[i].GetUint();
+        indices[i] = (unsigned short)indices_val_array[i].GetUint();
 
     meshdata->subMeshIndices.push_back(indices);
     meshdata->subMeshAABB.push_back(calculateAABB(meshdata->vertex, meshdata->getPerVertexSize(), indices));
@@ -1239,12 +1239,12 @@ bool Bundle3D::loadMeshDataJson_0_2(MeshDatas& meshdatas)
         unsigned int indexnum = mesh_submesh_val[INDEXNUM].GetUint();
 
         // indices
-        std::vector<unsigned int> indices;
+        std::vector<unsigned short> indices;
         indices.resize(indexnum);
 
         const rapidjson::Value& indices_val_array = mesh_submesh_val[INDICES];
         for (rapidjson::SizeType j = 0; j < indices_val_array.Size(); ++j)
-            indices[j] = (unsigned int)indices_val_array[j].GetUint();
+            indices[j] = (unsigned short)indices_val_array[j].GetUint();
 
         meshdata->subMeshIndices.push_back(indices);
         meshdata->subMeshAABB.push_back(calculateAABB(meshdata->vertex, meshdata->getPerVertexSize(), indices));
@@ -2324,7 +2324,7 @@ Bundle3D::~Bundle3D()
 
 cocos2d::AABB Bundle3D::calculateAABB(const std::vector<float>& vertex,
                                       int stride,
-                                      const std::vector<unsigned int>& index)
+                                      const std::vector<unsigned short>& index)
 {
     AABB aabb;
     stride /= 4;
